+++ CHANGES.md
@@ -9,6 +9,7 @@ Apollo 2.5.0
 * [Feature: Provide a new configfiles API to return the raw content of configuration files directly](https://github.com/apolloconfig/apollo/pull/5336)
 * [Feature: Enhanced instance configuration auditing and caching](https://github.com/apolloconfig/apollo/pull/5361)
 * [Feature: Provide a new open API to return the organization list](https://github.com/apolloconfig/apollo/pull/5365)
+* [Refactor: Exception handler adds root cause information](https://github.com/apolloconfig/apollo/pull/5367)
 
 ------------------
 All issues and pull requests are [here](https://github.com/apolloconfig/apollo/milestone/16?closed=1)